Shift question request back to question display
Fix history by letting the question display handle question request
